Read and parse .kml-files instead of .csv if necessary
This commit ensures that the MOSMIX-forecasts are read from the
.kml-files instead of .csv for requests concerning a time frame _after_
the support for .csv was dropped by DWD on 2018-09-17. In this commit,
the extraction of the .kml-file from the .kmz-file is not yet
implemented!
__Careful:__ since the object returned by `readTimeseriesDataMosmix()`
was changed, this __breaks__ `dwd_data_access` unless the file
`configuration/vois_data_access.json` is properly updated in the
repository of that microservice!